requirement.txt
fastapi
uvicorn
sqlalchemy
pymysql
dotenv
passlib[argon2]
python-jose[cryptography]
python-multipart
pydantic[email]
